Update the signature of allowEmptyString()
We made a mistake in the argument order of this method (and its
friends). The `message` argument is used far more frequently and should
have been first. With this change we can accept both argument orders and
emit deprecations when we get a `when` callable or bool as the second
argument.