Make signatures of allowEmpty* and notEmpty* methods consistent
Enable a backwards compatible shim for the allowEmpty* methods that
allow them to be called with `message, when` this allows us to normalize
the signatures with the newer notEmpty* methods and prioritize the more
frequently used argument (message) before a less frequently used when
condition.

The 'old' argument order can be removed in 5.0 as it is a low cost check
to have.